Specify ruby version
A number of ruby environment management tools support reading the ruby
version from a `.ruby-version` file, including rvm, rbenv, and chruby.
Meanwhile, bundler supports specifying a ruby version in the Gemfile and
Heroku honors that setting. Since Gemfiles are interpreted as ruby code,
set the variable in `.ruby-version` but read it in the Gemfile.